+++ neutron/db/migration/alembic_migrations/versions/f9263d6df56_remove_dhcp_lease.py
@@ -33,12 +33,33 @@ migration_for_plugins = [
     '*'
 ]
 
+from alembic import context
 from alembic import op
 import sqlalchemy as sa
 
+from neutron.db import migration
+
 
 def upgrade(active_plugins=None, options=None):
-    op.drop_column('ipallocations', u'expiration')
+
+    if migration.migration_engine(context) != 'sqlite':
+        op.drop_column('ipallocations', u'expiration')
+    else:
+        migration.rename_table('ipallocations', op)
+        migration.create_table(
+            'ipallocations',
+            sa.Column('port_id', sa.String(length=36), nullable=True),
+            sa.Column('ip_address', sa.String(length=64), nullable=False),
+            sa.Column('subnet_id', sa.String(length=36), nullable=False),
+            sa.Column('network_id', sa.String(length=36), nullable=False),
+            sa.ForeignKeyConstraint(['network_id'], ['networks.id'],
+                                    ondelete='CASCADE'),
+            sa.ForeignKeyConstraint(['port_id'], ['ports.id'],
+                                    ondelete='CASCADE'),
+            sa.ForeignKeyConstraint(['subnet_id'], ['subnets.id'],
+                                    ondelete='CASCADE'),
+            sa.PrimaryKeyConstraint('ip_address', 'subnet_id', 'network_id'),
+            op=op)
 
 
 def downgrade(active_plugins=None, options=None):
